fixed few imports and long to integer

+++ src/main/java/com/example/owetracker/repository/ExpenseRepository.java
@@ -14,9 +14,9 @@ import java.util.List;
 @Repository
 public interface ExpenseRepository extends JpaRepository<Expense, Long> {
 
-    List<Expense> findByGroupId(Long groupId);
+    List<Expense> findByGroupId(Integer groupId);
 
-    List<Expense> findByPaidById(Long userId);
+    List<Expense> findByPaidById(Integer userId);
 
     List<Expense> findByStatus(String status);
